requirements-test.txt
# Additional dependencies for local testing and development
# Install with: pip install -r requirements.txt -r requirements-test.txt

# Performance Testing & Monitoring
locust==2.20.0
django-debug-toolbar==4.2.0

# For test result analysis
matplotlib>=3.8.0
pandas>=2.0.0

# Testing
pytest>=8.0.0
pytest-django>=4.7.0
